Split option parser out of Options.hpp
Fixes #298
