+++ tests/Unit/Options/Test_Factory.cpp
@@ -9,6 +9,7 @@
 #include <vector>
 
 #include "Options/Options.hpp"
+#include "Options/ParseOptions.hpp"
 #include "tests/Unit/TestHelpers.hpp"
 
 namespace {
